+++ src/singletoneConstants.ts
@@ -14,10 +14,6 @@ export class Asterisk {
   public getStmt(): string {
     return '*'
   }
-
-  public toString(): string {
-    return this.getStmt()
-  }
 }
 
 export const ASTERISK = Asterisk.getInstance()
@@ -61,3 +57,23 @@ export class All {
 }
 
 export const ALL = All.getInstance()
+
+export class Default {
+  private static instance: Default
+  private readonly unique: symbol = Symbol()
+
+  private constructor() {}
+
+  public static getInstance(): Default {
+    if (!Default.instance) {
+      Default.instance = new Default()
+    }
+    return Default.instance
+  }
+
+  public getStmt(): string {
+    return 'DEFAULT'
+  }
+}
+
+export const DEFAULT = Default.getInstance()