+++ src/steps/BaseStep.ts
@@ -7,6 +7,9 @@ import { LogicalOperator } from '../operators'
 import { DeleteWithoutConditionError, TableNotFoundError } from '../errors'
 import { AliasedTable, Table } from '../database'
 import { FromItemInfo, FromItemRelation } from '../FromItemInfo'
+import { getStmtNull, getStmtBoolean, getStmtString, getStmtDate } from '../util'
+import { Binder } from '../binder'
+import { Default } from '../singletoneConstants'
 
 export enum Parenthesis {
   Open = '(',
@@ -25,7 +28,33 @@ export abstract class BaseStep {
   }
 
   private getStatement(): string {
-    let result = `${this.data.sqlPath}${this.data.distinct}`
+    let result = ''
+    switch (this.data.sqlPath) {
+    case SqlPath.SELECT:
+      result = this.getSelectStatement()
+      break
+    case SqlPath.DELETE:
+      result = this.getDeleteStatement()
+      break
+    case SqlPath.INSERT:
+      result = this.getInsertStatement()
+      break
+    case SqlPath.UPDATE:
+      result = this.getUpdateStatement()
+      break
+    }
+
+    if (this.data.option.useSemicolonAtTheEnd) result += ';'
+
+    return result
+  }
+
+  private getSelectStatement(): string {
+    let result = `SELECT`
+
+    if (this.data.distinct) {
+      result += ` ${this.data.distinct}`
+    }
 
     if (this.data.selectItemInfos.length > 0) {
       const selectPartsString = this.data.selectItemInfos.map(it => {
@@ -38,18 +67,7 @@ export abstract class BaseStep {
       result += ` FROM ${this.data.fromItemInfos.map(it => it.getStmt(this.data)).join('')}`
     }
 
-    if (this.data.whereParts.length > 0) {
-      BaseStep.throwIfConditionPartsInvalid(this.data.whereParts)
-      const wherePartsString = this.data.whereParts.map(it => {
-        if (it instanceof Condition || it instanceof Expression || it instanceof BooleanColumn) {
-          return it.getStmt(this.data)
-        }
-        return it.toString()
-      })
-      result += ` WHERE ${wherePartsString.join(' ')}`
-    } else if (this.data.sqlPath === SqlPath.DELETE && this.data.option.throwErrorIfDeleteHasNoCondition) {
-      throw new DeleteWithoutConditionError(`Delete statement must have where conditions or set throwErrorIfDeleteHasNoCondition option to false`)
-    }
+    result += this.getWhereParts()
 
     if (this.data.groupByItems.length > 0) {
       result += ` GROUP BY ${this.data.groupByItems.map(it => it.getStmt(this.data)).join(', ')}`
@@ -85,23 +103,114 @@ export abstract class BaseStep {
       result += ` OFFSET ${this.data.offset}`
     }
 
+    return result
+  }
+
+  private getDeleteStatement(): string {
+    let result = `DELETE`
+
+    if (this.data.fromItemInfos.length > 0) {
+      // todo: throw if fromItemInfos.length > 1
+      result += ` FROM ${this.data.fromItemInfos[0].getStmt(this.data)}`
+    }
+
+    result += this.getWhereParts()
+    result += this.getReturningParts()
+
+    return result
+  }
+
+  private getInsertStatement(): string {
+    let result = 'INSERT'
+    if (this.data.insertIntoTable !== undefined) {
+      result += ` INTO ${this.data.insertIntoTable.getStmt(this.data)}`
+      if (this.data.insertIntoColumns.length > 0) {
+        result += `(${this.data.insertIntoColumns.map(it => it.getDoubleQuotedName()).join(', ')})`
+      }
+      if (this.data.insertIntoValues.length > 0) {
+        const valueStringArray = this.data.insertIntoValues.map(it => {
+          if (it === null) {
+            return getStmtNull()
+          } else if (typeof it === 'boolean') {
+            return getStmtBoolean(it)
+          } else if (typeof it === 'number') {
+            return it.toString()
+          } else if (typeof it === 'string') {
+            return getStmtString(it)
+          } else if (it instanceof Date) {
+            return getStmtDate(it)
+          } else if (it instanceof Binder) {
+            if (it.no === undefined) {
+              this.data.binderStore.add(it)
+            }
+            return it.getStmt()
+          } else if (it instanceof Default) {
+            return it.getStmt()
+          } else {
+            throw new Error(`Value step has Unsupported value: ${it}, type: ${typeof it}`)
+          }
+        })
+        result += ` VALUES(${valueStringArray.join(', ')})`
+      } else if (this.data.insertIntoDefaultValues) {
+        result += ' DEFAULT VALUES'
+      } else if (this.data.selectItemInfos.length > 0) {
+        result += ` ${this.getSelectStatement()}`
+      } else {
+        throw new Error('Insert statement must have values or select items')
+      }
+    }
+
+    result += this.getReturningParts()
+
+    return result
+  }
+
+  private getUpdateStatement(): string {
+    let result = 'UPDATE'
+    if (this.data.updateTable !== undefined) {
+      result += ` ${this.data.updateTable.getStmt(this.data)}`
+      if (this.data.updateSetItemInfos.length > 0) {
+        result += ` SET ${this.data.updateSetItemInfos.map(it => it.getStmt(this.data)).join(', ')}`
+      }
+      result += this.getWhereParts()
+      result += this.getReturningParts()
+    }
+    return result
+  }
+
+  private getWhereParts(): string {
+    if (this.data.whereParts.length > 0) {
+      BaseStep.throwIfConditionPartsInvalid(this.data.whereParts)
+      const wherePartsString = this.data.whereParts.map(it => {
+        if (it instanceof Condition || it instanceof Expression || it instanceof BooleanColumn) {
+          return it.getStmt(this.data)
+        }
+        return it.toString()
+      })
+      return ` WHERE ${wherePartsString.join(' ')}`
+    }
+
+    if (this.data.sqlPath === SqlPath.DELETE && this.data.option.throwErrorIfDeleteHasNoCondition) {
+      throw new DeleteWithoutConditionError(`Delete statement must have where conditions or set throwErrorIfDeleteHasNoCondition option to false`)
+    }
+
+    return ''
+  }
+
+  private getReturningParts(): string {
     if (this.data.returning.length > 0) {
       const returningPartsString = this.data.returning.map(it => {
         return it.getStmt(this.data)
       })
-      result += ` RETURNING ${returningPartsString.join(', ')}`
+      return ` RETURNING ${returningPartsString.join(', ')}`
     }
-
-    if (this.data.option.useSemicolonAtTheEnd)
-      result += ';'
-
-    return result
+    return ''
   }
 
   public cleanUp() {
     this.data.sqlPath = undefined
     this.data.selectItemInfos.length = 0
-    this.data.distinct = ''
+    this.data.distinct = undefined
     this.data.fromItemInfos.length = 0
     this.data.whereParts.length = 0
     this.data.groupByItems.length = 0
@@ -109,6 +218,12 @@ export abstract class BaseStep {
     this.data.orderByItemInfos.length = 0
     this.data.limit = undefined
     this.data.offset = undefined
+    this.data.insertIntoTable = undefined
+    this.data.insertIntoColumns.length = 0
+    this.data.insertIntoValues.length = 0
+    this.data.insertIntoDefaultValues = false
+    this.data.updateTable = undefined
+    this.data.updateSetItemInfos.length = 0
     this.data.returning.length = 0
     this.data.binderStore.cleanUp()
   }