+++ src/util.ts
@@ -1,10 +1,10 @@
 import { Step } from './steps/Step'
 
-export function escapeDoubleQuote(source:string): string {
+export function escapeDoubleQuote(source: string): string {
   return source.replace(/"/g, '""')
 }
 
-export function escapeSingleQuote(source:string): string {
+function escapeSingleQuote(source: string): string {
   return source.replace(/'/g, "''")
 }
 
@@ -13,4 +13,20 @@ export function returnStepOrThrow(step: Step|undefined): Step {
     throw new Error('Step property in builder data is not initialized')
   }
   return step
+}
+
+export function getStmtNull(): string {
+  return 'NULL'
+}
+
+export function getStmtBoolean(value: boolean): string {
+  return value ? 'TRUE' : 'FALSE'
+}
+
+export function getStmtString(value: string): string {
+  return `'${escapeSingleQuote(value)}'`
+}
+
+export function getStmtDate(value: Date): string {
+  return `'${escapeSingleQuote(value.toISOString())}'`
 }