+++ test/nonExported/error.test.ts
@@ -1,15 +1,19 @@
 import {
   InvalidConditionError,
   Builder,
+  e,
 } from '../../src'
 
 // test non-exported Classes
 import { Condition } from '../../src/models/Condition'
 import { Expression } from '../../src/models/Expression'
 import { OnStep } from '../../src/steps/OnStep'
+import { BinderStore } from '../../src/binder'
+import { BuilderData } from '../../src/builder'
+import { ItemInfo } from '../../src/ItemInfo'
+import { Column } from '../../src/columns'
 
 import { database } from '../database'
-import { BinderStore } from '../../src/binder'
 
 //Alias
 const table1 = database.s.public.t.table1
@@ -33,33 +37,62 @@ describe('Throw desired Errors', () => {
   })
 
   describe('Error: "Step property in builder data is not initialized"', () => {
+    const data: BuilderData = {
+      binderStore: new BinderStore(),
+      database,
+      distinct: undefined,
+      fromItemInfos: [],
+      groupByItems: [],
+      havingParts: [],
+      option: {
+        useSemicolonAtTheEnd: true,
+        addAscAfterOrderByItem: 'when mentioned',
+        addNullsLastAfterOrderByItem: 'when mentioned',
+        addAsBeforeColumnAlias: 'always',
+        addPublicSchemaName: 'never',
+        addTableName: 'when two tables or more',
+        addAsBeforeTableAlias: 'always',
+        throwErrorIfDeleteHasNoCondition: true,
+      },
+      orderByItemInfos: [],
+      selectItemInfos: [],
+      whereParts: [],
+      insertIntoTable: undefined,
+      insertIntoColumns: [],
+      insertIntoValues: [],
+      insertIntoDefaultValues: false,
+      updateTable: undefined,
+      updateSetItemInfos: [],
+      returning: [],
+    }
     it(`Throws error when Step is not initialized`, () => {
       function actual() {
-        new OnStep({
-          binderStore: new BinderStore(),
-          database,
-          distinct: '',
-          fromItemInfos: [],
-          groupByItems: [],
-          havingParts: [],
-          option: {
-            useSemicolonAtTheEnd: true,
-            addAscAfterOrderByItem: 'when mentioned',
-            addNullsLastAfterOrderByItem: 'when mentioned',
-            addAsBeforeColumnAlias: 'always',
-            addPublicSchemaName: 'never',
-            addTableName: 'when two tables or more',
-            addAsBeforeTableAlias: 'always',
-            throwErrorIfDeleteHasNoCondition: true,
-          },
-          orderByItemInfos: [],
-          selectItemInfos: [],
-          whereParts: [],
-          returning: [],
-        }).crossJoin(table1)
+        new OnStep(data).crossJoin(table1)
       }
 
       expect(actual).toThrow(`Step property in builder data is not initialized`)
     })
+  })
+
+  it(`Throws: ItemInfo is an abstract class`, () => {
+    class DummyItemInfo extends ItemInfo {
+      constructor() {super()}
+      getColumns(): Column[] {
+        return []
+      }
+
+      getStmt(data: BuilderData): string {
+        return ''
+      }
+    }
+    function actual() {
+      sql
+        .deleteFrom(table1)
+        .where(e(1).eq(1))
+        .returning(new DummyItemInfo())
+        .getSQL()
+    }
+
+    expect(actual).toThrow(`ItemInfo is an abstract class`)
   })
 })